backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pypdf==3.10.0
chromadb==0.4.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.6
alembic==1.11.1
